Refactor and Doc: New Inheritance structure for less code repetition and docs for all the object methods
