+++ net-sdk/src/internal_classes/SerieAbstract.cs
@@ -1,0 +1,34 @@
+using System;
+
+namespace net_sdk.src.internal_classes;
+
+public abstract record class SerieAbstract(
+    string Id,
+    string Name,
+    string? Logo
+
+) : Model()
+{
+    /// <summary>
+    /// Returns the URL of the logo of the serie. If the serie has no logo, null is returned.
+    /// </summary>
+    /// <param name="extension"></param>
+    /// <returns></returns>
+    public string? GetLogoUrl(Extension extension)
+    {
+        if (Logo == null) return null;
+        return $"{Logo}.{extension}";
+    }
+    /// <summary>
+    /// Async returns the logo of the serie as a byte array. If the serie has no logo, null is returned.
+    /// </summary>
+    /// <param name="extension"></param>
+    /// <returns></returns>
+    public async Task<byte[]?> GetLogo(Extension extension)
+    {
+        var logoUrl = GetLogoUrl(extension);
+        if (logoUrl == null) return null;
+        return await TCGDex.GetImage(logoUrl);
+        
+    }
+}
