+++ net-sdk/src/models/CardResume.cs
@@ -8,22 +8,14 @@ public record class CardResume(
     string LocalId,
     string Name,
     string? Image
-) : Model()
+) : CardAbstract(Id, LocalId, Name, Image)
 {
-    public string GetImageUrl(Quality quality, Extension extension)
-    {
-        return $"{this.Image}/{quality}.{extension}";
-    }
-
-    public async Task<byte[]?> GetImage(Quality quality, Extension extension)
-    {
-        
-        return await TCGDex.GetImage(GetImageUrl(quality, extension));
-    }
-
+    /// <summary>
+    /// Async returns a <see cref="Card"/> of the CardResume.
+    /// </summary>
+    /// <returns></returns>
     public async Task<Card?> GetFullCard()
     {
-        //tCGDex = new TCGDex("en");
-        return await this.TCGDex.FetchCard(Id);
+        return await TCGDex.FetchCard(Id);
     }
 }
