+++ net-sdk/src/models/SetResume.cs
@@ -9,52 +9,8 @@ public record class SetResume(
     string? Logo,
     string? Symbol,
     SetCardCount CardCount
-) : Model()
+) : SetAbstract(Id, Name, Logo, Symbol, CardCount)
 {
-    /// <summary>
-    /// Async returns the logo url of the set, based on the <see cref="Logo"/> property and the given <paramref name="extension"/>.
-    /// </summary>
-    /// <param name="extension"></param>
-    /// <returns>Returns the logo url of the set as a string.</returns>
-    public string? GetLogoUrl(Extension extension)
-    {
-        if (Logo == null) return null;
-        return $"{Logo}.{extension}";
-    }
-    /// <summary>
-    /// Async returns the logo of the set, based on the <see cref="Logo"/> property and the given <paramref name="extension"/>.
-    /// </summary>
-    /// <param name="extension"></param>
-    /// <returns>Returns the logo of the set as a byte array.</returns>
-    public async Task<byte[]?> GetLogo(Extension extension)
-    {
-        var logoUrl = GetLogoUrl(extension);
-        if (logoUrl == null) return null;
-        return await TCGDex.GetImage(logoUrl);
-        
-    }
-    /// <summary>
-    /// Async returns the symbol url of the set, based on the <see cref="Symbol"/> property and the given <paramref name="extension"/>.
-    /// </summary>
-    /// <param name="extension"></param>
-    /// <returns>Returns the symbol url of the set as a string.</returns>
-    public string? GetSymbolUrl(Extension extension)
-    {
-        if (Symbol == null) return null;
-        return $"{Symbol}.{extension}";
-    }
-    /// <summary>
-    /// Async returns the symbol of the set, based on the <see cref="Symbol"/> property and the given <paramref name="extension"/>.
-    /// </summary>
-    /// <param name="extension"></param>
-    /// <returns>Returns the symbol of the set as a byte array.</returns>
-    public async Task<byte[]?> GetSymbol(Extension extension)
-    {
-        var symbolUrl = GetSymbolUrl(extension);
-        if (symbolUrl == null) return null;
-        return await TCGDex.GetImage(symbolUrl);
-        
-    }
     /// <summary>
     /// Async returns a <see cref="Set"/> of the SetResume.
     /// </summary>
